# sprint_3/task2/device-management-api/src/main/java/ru/yandex/practcum/device_management_api/model/rest/DeviceActionResponse.java
package ru.yandex.practcum.device_management_api.model.rest;

import lombok.AllArgsConstructor;
import lombok.Data;

@Data
@AllArgsConstructor
public class DeviceActionResponse {
    private String status;
}