# sprint_3/task2/device-management-api/src/main/java/ru/yandex/practcum/device_management_api/model/rest/DeviceTelemetryResponse.java
package ru.yandex.practcum.device_management_api.model.rest;

import lombok.Builder;
import lombok.Data;

import java.time.LocalDateTime;

@Data
@Builder(toBuilder = true)
public class DeviceTelemetryResponse {
    private String status;
    private Long accountId;
    private String serialNumber;
    private LocalDateTime updatedAt;
    private String data;
}